Zod Setup Completed and Tested
Left a Test Component for you to check how zod works with a form, you can place the test component in the App.tsx file and run it. I also made a little form schema, to be upgraded in the future when needed. FYI, we will need to set an exact schema for every input we need, so please refer to the zod documentation when needed. Note: I installed react-hook-form and @hookform/resolvers to link zod to a from.